refactor: colocate markdown rendering logic
