--- src/remark.ts
@@ -1,6 +1,0 @@
-import { remark as createRemark } from 'remark'
-import gfm from 'remark-gfm'
-
-export const remark = createRemark().use(gfm).data('settings', {
-  bullet: '-',
-})
